Simplify class-hierarchy of client-side Cache
In this commit none of the functionality is added or removed.

In order to simplify classes of Cache the following steps are done:

1. ClientCacheProxy contains mostly all overrides from ICacheInternal.
All methods are tend to be without deep low-level code.
2. ClientCacheProxySupport is only one superclass of ClientCacheProxy
(before ClientProxy level) that contains internal re-used methods and
low-level code.
3. Utility-methods and utility-classes are moved to
ClientCacheProxySupportUtil class.
4. CountDownLatches and all related methods are moved to a standalone
class ClientCacheProxySyncListenerCompleter for logical isolation and
complexity reasons.
5. Cache entry listener registrations and all related methods are moved
to a standalone class ClientCacheProxySyncListenerCompleter.
